+++ sbe-tool/src/test/cpp/Rc3OtfFullIrTest.cpp
@@ -249,17 +249,17 @@ public:
             .next().speed(fuel1Speed).mpg(fuel1Mpg);
 
         fuelFigures.putUsageDescription(
-            FUEL_FIGURES_1_USAGE_DESCRIPTION, static_cast<int>(strlen(FUEL_FIGURES_1_USAGE_DESCRIPTION)));
+            FUEL_FIGURES_1_USAGE_DESCRIPTION, static_cast<std::uint16_t>(strlen(FUEL_FIGURES_1_USAGE_DESCRIPTION)));
 
         fuelFigures
             .next().speed(fuel2Speed).mpg(fuel2Mpg);
         fuelFigures.putUsageDescription(
-            FUEL_FIGURES_2_USAGE_DESCRIPTION, static_cast<int>(strlen(FUEL_FIGURES_2_USAGE_DESCRIPTION)));
+            FUEL_FIGURES_2_USAGE_DESCRIPTION, static_cast<std::uint16_t>(strlen(FUEL_FIGURES_2_USAGE_DESCRIPTION)));
 
         fuelFigures
             .next().speed(fuel3Speed).mpg(fuel3Mpg);
         fuelFigures.putUsageDescription(
-            FUEL_FIGURES_3_USAGE_DESCRIPTION, static_cast<int>(strlen(FUEL_FIGURES_3_USAGE_DESCRIPTION)));
+            FUEL_FIGURES_3_USAGE_DESCRIPTION, static_cast<std::uint16_t>(strlen(FUEL_FIGURES_3_USAGE_DESCRIPTION)));
 
         Car::PerformanceFigures &perfFigs = car.performanceFiguresCount(PERFORMANCE_FIGURES_COUNT);
 
@@ -277,10 +277,10 @@ public:
             .next().mph(perf2bMph).seconds(perf2bSeconds)
             .next().mph(perf2cMph).seconds(perf2cSeconds);
 
-        car.putManufacturer(MANUFACTURER, static_cast<int>(strlen(MANUFACTURER)));
-        car.putModel(MODEL, static_cast<int>(strlen(MODEL)));
-        car.putActivationCode(ACTIVATION_CODE, static_cast<int>(strlen(ACTIVATION_CODE)));
-        car.putColor(COLOR, static_cast<int>(strlen(COLOR)));
+        car.putManufacturer(MANUFACTURER, static_cast<std::uint16_t>(strlen(MANUFACTURER)));
+        car.putModel(MODEL, static_cast<std::uint16_t>(strlen(MODEL)));
+        car.putActivationCode(ACTIVATION_CODE, static_cast<std::uint16_t>(strlen(ACTIVATION_CODE)));
+        car.putColor(COLOR, static_cast<std::uint32_t>(strlen(COLOR)));
 
         return hdr.encodedLength() + car.encodedLength();
     }
@@ -1055,7 +1055,7 @@ public:
             {
                 EXPECT_EQ(fieldToken.fieldId(), fieldIdFuelUsageDescription);
                 EXPECT_EQ(length, strlen(FUEL_FIGURES_1_USAGE_DESCRIPTION));
-                EXPECT_EQ(std::string(buffer, length), FUEL_FIGURES_1_USAGE_DESCRIPTION);
+                EXPECT_EQ(std::string(buffer, static_cast<std::size_t>(length)), FUEL_FIGURES_1_USAGE_DESCRIPTION);
                 break;
             }
 
@@ -1063,7 +1063,7 @@ public:
             {
                 EXPECT_EQ(fieldToken.fieldId(), fieldIdFuelUsageDescription);
                 EXPECT_EQ(length, strlen(FUEL_FIGURES_2_USAGE_DESCRIPTION));
-                EXPECT_EQ(std::string(buffer, length), FUEL_FIGURES_2_USAGE_DESCRIPTION);
+                EXPECT_EQ(std::string(buffer, static_cast<std::size_t>(length)), FUEL_FIGURES_2_USAGE_DESCRIPTION);
                 break;
             }
 
@@ -1071,7 +1071,7 @@ public:
             {
                 EXPECT_EQ(fieldToken.fieldId(), fieldIdFuelUsageDescription);
                 EXPECT_EQ(length, strlen(FUEL_FIGURES_3_USAGE_DESCRIPTION));
-                EXPECT_EQ(std::string(buffer, length), FUEL_FIGURES_3_USAGE_DESCRIPTION);
+                EXPECT_EQ(std::string(buffer, static_cast<std::size_t>(length)), FUEL_FIGURES_3_USAGE_DESCRIPTION);
                 break;
             }
 
@@ -1155,12 +1155,12 @@ TEST_F(Rc3OtfFullIrTest, shouldHandleAllEventsCorrectlyAndInOrder)
 
     EXPECT_EQ(headerDecoder.encodedLength(), MessageHeader::encodedLength());
     const char *messageBuffer = m_buffer + headerDecoder.encodedLength();
-    std::size_t length = encodedCarAndHdrLength - headerDecoder.encodedLength();
+    std::size_t length = static_cast<std::size_t>(encodedCarAndHdrLength - headerDecoder.encodedLength());
     std::uint64_t actingVersion = headerDecoder.getSchemaVersion(m_buffer);
     std::uint64_t blockLength = headerDecoder.getBlockLength(m_buffer);
 
     const std::size_t result = OtfMessageDecoder::decode(
-        messageBuffer, length, actingVersion, blockLength, messageTokens, *this);
+        messageBuffer, length, actingVersion, static_cast<std::size_t>(blockLength), messageTokens, *this);
     EXPECT_EQ(result, static_cast<std::size_t>(encodedCarAndHdrLength - MessageHeader::encodedLength()));
 }
 
@@ -1191,7 +1191,7 @@ TEST_P(Rc3OtfFullIrLengthTest, shouldExceptionIfLengthTooShort)
         std::unique_ptr<char[]> decodeBuffer(new char[length]);
 
         ::memcpy(decodeBuffer.get(), m_buffer + headerDecoder.encodedLength(), length);
-        OtfMessageDecoder::decode(decodeBuffer.get(), length, actingVersion, blockLength, messageTokens, *this);
+        OtfMessageDecoder::decode(decodeBuffer.get(), length, actingVersion, static_cast<std::size_t>(blockLength), messageTokens, *this);
     }, std::runtime_error);
 }
